services/minicpm/pyproject.toml
[project]
name = "minicpm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "Pillow==10.1.0",
    "torch==2.1.2",
    "torchvision==0.16.2",
    "transformers==4.40.0",
    "sentencepiece==0.1.99",
    "numpy==1.24.1",
]
